refactor: move module parsers to seperate directory
