+++ reports/jest-junit.xml
@@ -1,7 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testsuites name="jest tests" tests="20" failures="0" errors="0" time="18.141">
-  <testsuite name="src\extractors\youtube\proto\__tests__\protoTest.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T06:05:20" time="3.326" tests="8">
-    <testcase classname="protobuf parsing tests › if encodeVisitorData works" name="abcd and 100 should be CgRhYmNkKGQ%3D" time="0.01">
+<testsuites name="jest tests" tests="20" failures="0" errors="0" time="16.044">
+  <testsuite name="src\extractors\youtube\parsers\__tests__\parserTest.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T10:08:43" time="2.737" tests="4">
+    <testcase classname="parser error handling › videoDetail" name="if parser throws error on bad input" time="0.014">
+    </testcase>
+    <testcase classname="parser error handling › homePage" name="if parser throws error on bad input" time="0.002">
+    </testcase>
+    <testcase classname="parser error handling › bad" name="if parser throws error on bad input" time="0">
+    </testcase>
+    <testcase classname="parser handling" name="searchSuggestions" time="0.001">
+    </testcase>
+  </testsuite>
+  <testsuite name="src\extractors\youtube\proto\__tests__\protoTest.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T10:08:43" time="2.754" tests="8">
+    <testcase classname="protobuf parsing tests › if encodeVisitorData works" name="abcd and 100 should be CgRhYmNkKGQ%3D" time="0.006">
     </testcase>
     <testcase classname="protobuf parsing tests › if encodeVisitorData works" name="efgh and 200 should be CgRlZmdoKMgB" time="0">
     </testcase>
@@ -12,13 +22,13 @@
   order: &apos;rating&apos;,
   type: &apos;video&apos;,
   duration: &apos;medium&apos;
-} should be CAESBggFEAEYAw%3D%3D " time="0.005">
+} should be CAESBggFEAEYAw%3D%3D " time="0.003">
     </testcase>
-    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ type: &apos;channel&apos; } should be EgIQAg%3D%3D " time="0.001">
+    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ type: &apos;channel&apos; } should be EgIQAg%3D%3D " time="0">
     </testcase>
-    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ features: [Array] } should be EgQgASgB " time="0">
+    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ features: [Array] } should be EgQgASgB " time="0.001">
     </testcase>
-    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ features: [Array] } should be EgVAAbgBAQ%3D%3D " time="0.001">
+    <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{ features: [Array] } should be EgVAAbgBAQ%3D%3D " time="0">
     </testcase>
     <testcase classname="protobuf parsing tests › if encodeSearchFilter works" name="{
   uploadDate: &apos;week&apos;,
@@ -26,35 +36,25 @@
   duration: &apos;long&apos;,
   features: [Array],
   order: &apos;uploadDate&apos;
-} should be CAISCAgDEAEYAjgB " time="0">
+} should be CAISCAgDEAEYAjgB " time="0.004">
     </testcase>
   </testsuite>
-  <testsuite name="src\extractors\youtube\parsers\__tests__\parserTest.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T06:05:20" time="3.373" tests="4">
-    <testcase classname="parser error handling › videoDetail" name="if parser throws error on bad input" time="0.018">
+  <testsuite name="tests\main.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T10:08:43" time="15.275" tests="8">
+    <testcase classname="YouTube test suite › initilization tests" name="if init can handle error during initialization" time="0.011">
     </testcase>
-    <testcase classname="parser error handling › homePage" name="if parser throws error on bad input" time="0.003">
+    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="2.57">
     </testcase>
-    <testcase classname="parser error handling › bad" name="if parser throws error on bad input" time="0.005">
+    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="2.139">
     </testcase>
-    <testcase classname="parser handling" name="searchSuggestions" time="0.002">
+    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="2.763">
     </testcase>
-  </testsuite>
-  <testsuite name="tests\main.test.ts" errors="0" failures="0" skipped="0" timestamp="2022-06-08T06:05:20" time="17.528" tests="8">
-    <testcase classname="YouTube test suite › initilization tests" name="if init can handle error during initialization" time="0.012">
+    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo handles errors correctly" time="0.821">
     </testcase>
-    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="2.764">
+    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo handles errors correctly" time="0.899">
     </testcase>
-    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="2.255">
+    <testcase classname="YouTube test suite" name="if home page can be fetched" time="2.159">
     </testcase>
-    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo works as expected" time="3.165">
-    </testcase>
-    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo handles errors correctly" time="1.105">
-    </testcase>
-    <testcase classname="YouTube test suite › getVideoInfo tests" name="if getVideoInfo handles errors correctly" time="1.128">
-    </testcase>
-    <testcase classname="YouTube test suite" name="if home page can be fetched" time="2.361">
-    </testcase>
-    <testcase classname="YouTube test suite" name="if search suggestions can be fetched" time="0.606">
+    <testcase classname="YouTube test suite" name="if search suggestions can be fetched" time="0.399">
     </testcase>
   </testsuite>
 </testsuites>
